etl_case_status/requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9


